# apps/api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OptiFlow AI API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devis": {
      "post": {
        "tags": [
          "devis"
        ],
        "summary": "Create Devis",
        "operationId": "create_devis_api_v1_devis_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevisCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "devis"
        ],
        "summary": "List Devis",
        "operationId": "list_devis_api_v1_devis_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DevisResponse"
                  },
                  "title": "Response List Devis Api V1 Devis Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devis/{devis_id}": {
      "get": {
        "tags": [
          "devis"
        ],
        "summary": "Get Devis",
        "operationId": "get_devis_api_v1_devis__devis_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "devis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Devis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevisDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "devis"
        ],
        "summary": "Update Devis",
        "operationId": "update_devis_api_v1_devis__devis_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "devis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Devis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevisUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devis/{devis_id}/status": {
      "patch": {
        "tags": [
          "devis"
        ],
        "summary": "Change Status",
        "operationId": "change_status_api_v1_devis__devis_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "devis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Devis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevisStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/gdpr/clients/{client_id}/data": {
      "get": {
        "tags": [
          "gdpr"
        ],
        "summary": "Get Client Data",
        "operationId": "get_client_data_api_v1_gdpr_clients__client_id__data_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Client Data Api V1 Gdpr Clients  Client Id  Data Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gdpr/clients/{client_id}/export": {
      "post": {
        "tags": [
          "gdpr"
        ],
        "summary": "Export Client Data",
        "operationId": "export_client_data_api_v1_gdpr_clients__client_id__export_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/gdpr/clients/{client_id}/anonymize": {
      "post": {
        "tags": [
          "gdpr"
        ],
        "summary": "Anonymize Client",
        "operationId": "anonymize_client_api_v1_gdpr_clients__client_id__anonymize_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "client_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Anonymize Client Api V1 Gdpr Clients  Client Id  Anonymize Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DevisCreate": {
        "properties": {
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "part_secu": {
            "type": "number",
            "minimum": 0.0,
            "title": "Part Secu",
            "default": 0
          },
          "part_mutuelle": {
            "type": "number",
            "minimum": 0.0,
            "title": "Part Mutuelle",
            "default": 0
          },
          "lignes": {
            "items": {
              "$ref": "#/components/schemas/DevisLineCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Lignes"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "lignes"
        ],
        "title": "DevisCreate"
      },
      "DevisDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "montant_ht": {
            "type": "number",
            "title": "Montant Ht"
          },
          "tva": {
            "type": "number",
            "title": "Tva"
          },
          "montant_ttc": {
            "type": "number",
            "title": "Montant Ttc"
          },
          "part_secu": {
            "type": "number",
            "title": "Part Secu"
          },
          "part_mutuelle": {
            "type": "number",
            "title": "Part Mutuelle"
          },
          "reste_a_charge": {
            "type": "number",
            "title": "Reste A Charge"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "lignes": {
            "items": {
              "$ref": "#/components/schemas/DevisLineResponse"
            },
            "type": "array",
            "title": "Lignes",
            "default": []
          },
          "customer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "numero",
          "status",
          "montant_ht",
          "tva",
          "montant_ttc",
          "part_secu",
          "part_mutuelle",
          "reste_a_charge",
          "created_at"
        ],
        "title": "DevisDetail"
      },
      "DevisLineCreate": {
        "properties": {
          "designation": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Designation"
          },
          "quantite": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantite",
            "default": 1
          },
          "prix_unitaire_ht": {
            "type": "number",
            "minimum": 0.0,
            "title": "Prix Unitaire Ht"
          },
          "taux_tva": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Taux Tva",
            "default": 20.0
          }
        },
        "type": "object",
        "required": [
          "designation",
          "prix_unitaire_ht"
        ],
        "title": "DevisLineCreate"
      },
      "DevisLineResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "designation": {
            "type": "string",
            "title": "Designation"
          },
          "quantite": {
            "type": "integer",
            "title": "Quantite"
          },
          "prix_unitaire_ht": {
            "type": "number",
            "title": "Prix Unitaire Ht"
          },
          "taux_tva": {
            "type": "number",
            "title": "Taux Tva"
          },
          "montant_ht": {
            "type": "number",
            "title": "Montant Ht"
          },
          "montant_ttc": {
            "type": "number",
            "title": "Montant Ttc"
          }
        },
        "type": "object",
        "required": [
          "id",
          "designation",
          "quantite",
          "prix_unitaire_ht",
          "taux_tva",
          "montant_ht",
          "montant_ttc"
        ],
        "title": "DevisLineResponse"
      },
      "DevisResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "numero": {
            "type": "string",
            "title": "Numero"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "montant_ht": {
            "type": "number",
            "title": "Montant Ht"
          },
          "tva": {
            "type": "number",
            "title": "Tva"
          },
          "montant_ttc": {
            "type": "number",
            "title": "Montant Ttc"
          },
          "part_secu": {
            "type": "number",
            "title": "Part Secu"
          },
          "part_mutuelle": {
            "type": "number",
            "title": "Part Mutuelle"
          },
          "reste_a_charge": {
            "type": "number",
            "title": "Reste A Charge"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_id",
          "numero",
          "status",
          "montant_ht",
          "tva",
          "montant_ttc",
          "part_secu",
          "part_mutuelle",
          "reste_a_charge",
          "created_at"
        ],
        "title": "DevisResponse"
      },
      "DevisStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(brouillon|envoye|signe|facture|annule)$",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DevisStatusUpdate"
      },
...
      "DevisUpdate": {
        "properties": {
          "part_secu": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Secu"
          },
          "part_mutuelle": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Mutuelle"
          },
          "lignes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DevisLineCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lignes"
          }
        },
        "type": "object",
        "title": "DevisUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}